refactor(e2e): extract shared video recording rig into @nteract/e2e-record
apps/notebook and apps/notebook-cloud each carried a near-verbatim copy of the
recording rig: show-video.mjs, show-frames.mjs, test-results.mjs,
video-viewer.html, and ~120 lines of helpers.ts. The only real differences were
root paths, usage strings, and the recorded viewport size.

Move the mechanics to packages/e2e-record. Each app now keeps a thin shim
(show-video.mjs, show-frames.mjs, recording-paths.mjs) naming its own
test-results/, recordings/, and record command, plus the app-specific readiness
marks in helpers.ts.

Couple the zoom to the viewport. The 2x CSS zoom lived in helpers.ts while the
doubled viewport lived in each playwright.config.ts, so changing one without the
other silently mis-framed the video. Configs now declare the effective CSS
framing they want -- videoRecordingUse({ width: 600, height: 400 }) -- and both
the capture size and the page zoom derive from one VIDEO_ZOOM constant.

Unify the viewer. The cloud copy sized responsively and its comment claimed the
cloud recorder captured 1:1 CSS pixels, but cloud helpers.ts applied the same 2x
zoom -- the divergence was accidental. The shared viewer scales by
devicePixelRatio for physical-pixel accuracy and caps to the window via
aspect-ratio so an oversized clip still fits.

Verified end to end: recorded add-cell (1200x800 = 600x400 x2) and the cloud
home-to-notebook flow (2400x1600 = 1200x800 x2), both matching their prior
sizes; show-frames and show-video produce frames, trim metadata, nested archive
dirs, and a working viewer for each app. Non-recording runs still emit no
artifacts.
