+++ .agents/skills/record-ui/SKILL.md
@@ -20,6 +20,11 @@ pacing pauses, and screenshot capture. Without it a spec runs as a normal test:
 no video, no PNGs, no pacing delay. So a spec with `screenshot()` calls is safe
 to commit and run in CI.
 
+**Two apps, one rig.** The mechanics live in `packages/e2e-record` and are shared
+by `apps/notebook` (desktop) and `apps/notebook-cloud` (hosted). This page uses
+the desktop paths; for cloud flows the commands live under `e2e/browser/` and the
+runner is `run-cloud-e2e.mjs` — see the [cloud section](#recording-a-cloud-flow).
+
 ## Agent self-check loop
 
 1. Write a spec that drives the change; drop `await screenshot(page, "label")`
@@ -76,16 +81,28 @@ test.describe("<feature>", () => {
 
 ## Recording resolution
 
-`playwright.config.ts` records at a 1200x800 viewport and `helpers.ts` CSS-zooms
-the page 2x, so the video frames the content a compact 600x400 window would
-show, at full retina density. This is deliberate: Playwright's recorder captures
-the viewport at CSS-pixel size and **ignores `deviceScaleFactor`**, so there is
-no "record at 2x" option — a 600px viewport yields a blurry 600px video. Change
-the size in one place (`videoSize` in the config), never per spec.
+Each app's `playwright.config.ts` declares the **effective CSS framing** it wants:
 
-`show-video.mjs` opens the result through `video-viewer.html`, which scales the
-video down by `devicePixelRatio` so it displays at the intended compact size
-instead of QuickTime's oversized raw pixels.
+```ts
+use: { ...videoRecordingUse({ width: 600, height: 400 }) }   // apps/notebook
+use: { ...videoRecordingUse({ width: 1200, height: 800 }) }  // apps/notebook-cloud
+```
+
+`videoRecordingUse()` multiplies that by `VIDEO_ZOOM` (2) to get the capture
+viewport, and `applyVideoZoom()` CSS-zooms the page by the same constant — so the
+clip shows exactly the requested framing at full retina density. Both read one
+constant, so they can't drift out of sync. This dance is necessary because
+Playwright's recorder captures the viewport at CSS-pixel size and **ignores
+`deviceScaleFactor`**: there is no "record at 2x" option, and a 600px viewport
+yields a blurry 600px video.
+
+Change the framing in the app's config, never per spec. Cloud records wider
+because the list/dashboard chrome clips at the desktop's compact size.
+
+`show-video.mjs` opens the result through the shared `video-viewer.html`, which
+scales the video down by `devicePixelRatio` so it displays at the intended size
+instead of QuickTime's oversized raw pixels (capped to the window, so an
+oversized clip still fits).
 
 ## Trimming the load prefix
 
@@ -172,14 +189,48 @@ showing up, check the flag is `true` in:
 - Nightly: `~/Library/Application Support/nteract-nightly/settings.json`
 - Stable:  `~/Library/Application Support/nteract/settings.json`
 
+## Recording a cloud flow
+
+Same rig, cloud paths. Run from `apps/notebook-cloud`:
+
+```bash
+NTERACT_E2E_RECORD=1 node e2e/browser/run-cloud-e2e.mjs <spec-file>
+node e2e/browser/show-frames.mjs [pattern]
+node e2e/browser/show-video.mjs [pattern]
+```
+
+Cloud specs import from `e2e/browser/helpers.ts` and start with
+`openHomePage(page)` (which marks the clip start itself, since the home-page
+render *is* the start of the footage), then `signInWithLocalAuth(page)`. See
+`apps/notebook-cloud/e2e/browser/home-to-notebook.spec.ts`.
+
 ## Recording helpers (e2e/helpers.ts)
+
+Each app re-exports the shared helpers from `@nteract/e2e-record` through its own
+`helpers.ts`, so specs always import from `./helpers`:
 
 | Helper | Purpose |
 |--------|---------|
-| `markKernelReady(page)` | Start the clip once the kernel is idle |
-| `markClipStart()` | Raw primitive: start the clip now, asserting nothing |
 | `screenshot(page, label)` | Capture `frame-NN-<label>.png` for agent inspection |
 | `pauseForVideo(page, ms)` | Sleep only while recording, to pace the video |
+| `markClipStart()` | Raw primitive: start the clip now, asserting nothing |
+| `markKernelReady(page)` | *(desktop)* Start the clip once the kernel is idle |
 
 The rest of `helpers.ts` (`openNotebookRoom`, `executeCell`,
 `waitForOutputContaining`, …) is the normal e2e surface — read the file.
+
+## Where the rig lives
+
+`packages/e2e-record` holds everything mechanical, shared by both apps:
+
+| Path | Purpose |
+|------|---------|
+| `src/recording.ts` | `screenshot`, `pauseForVideo`, `markClipStart`, zoom + `VIDEO_ZOOM` |
+| `src/config.ts` | `videoRecordingUse()` for each app's `playwright.config.ts` |
+| `src/show-video.mjs` | archive + ffmpeg encode + open the viewer |
+| `src/show-frames.mjs` | list a run's PNGs |
+| `src/video-viewer.html` | retina-correct playback page |
+
+Each app keeps only a thin `e2e/show-video.mjs` / `show-frames.mjs` shim plus a
+`recording-paths.mjs` naming its own `test-results/`, `recordings/`, and record
+command. Fix recording behavior in the package, not in an app.
