--- apps/notebook/e2e/video-viewer.html
@@ -1,95 +1,0 @@
-<!doctype html>
-<!--
-  Template for the recording viewer. show-video.mjs copies this next to the
-  encoded .webm in e2e/recordings/ and substitutes the placeholder below with
-  the video's filename, so the page references it by *relative* path.
-
-  Why not a ?v= query param: macOS `open` treats a file:// URL as a path and
-  silently strips the query string, so the page would always load with no video.
--->
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>e2e video viewer</title>
-    <style>
-      :root {
-        color-scheme: light dark;
-      }
-      html,
-      body {
-        margin: 0;
-        height: 100%;
-      }
-      body {
-        display: grid;
-        place-items: center;
-        gap: 12px;
-        background: #111;
-        color: #ddd;
-        font:
-          13px/1.4 ui-monospace,
-          SFMono-Regular,
-          Menlo,
-          monospace;
-        padding: 24px;
-        box-sizing: border-box;
-      }
-      /*
-       * The video is captured at 2x device pixels (e.g. 1200x800) so it stays
-       * crisp on retina. Displaying it at its intrinsic CSS size would be twice
-       * as large as intended, so we scale it down to intrinsicWidth / dpr:
-       * on a 2x display that maps one video pixel to one physical pixel — the
-       * compact framing the spec designed for, at full sharpness.
-       */
-      video {
-        display: block;
-        background: #000;
-        box-shadow: 0 8px 40px rgba(0, 0, 0, 0.6);
-        image-rendering: -webkit-optimize-contrast;
-      }
-      #meta {
-        max-width: 90vw;
-        word-break: break-all;
-        opacity: 0.7;
-        text-align: center;
-      }
-      #error {
-        color: #ff8a80;
-        max-width: 80vw;
-        text-align: center;
-      }
-    </style>
-  </head>
-  <body>
-    <video id="video" controls autoplay loop playsinline></video>
-    <div id="meta"></div>
-    <div id="error"></div>
-
-    <script>
-      const VIDEO_FILE = "__VIDEO_FILE__";
-
-      const video = document.getElementById("video");
-      const meta = document.getElementById("meta");
-      const errorEl = document.getElementById("error");
-
-      video.src = encodeURIComponent(VIDEO_FILE);
-
-      const dpr = window.devicePixelRatio || 1;
-
-      video.addEventListener("loadedmetadata", () => {
-        const cssWidth = video.videoWidth / dpr;
-        const cssHeight = video.videoHeight / dpr;
-        video.style.width = cssWidth + "px";
-        video.style.height = cssHeight + "px";
-        meta.innerHTML =
-          `${video.videoWidth}×${video.videoHeight} intrinsic → ` +
-          `${Math.round(cssWidth)}×${Math.round(cssHeight)} CSS (dpr ${dpr})<br>${VIDEO_FILE}`;
-      });
-
-      video.addEventListener("error", () => {
-        errorEl.textContent = "Failed to load the video: " + VIDEO_FILE;
-      });
-    </script>
-  </body>
-</html>
