+++ packages/e2e-record/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@nteract/e2e-record",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./config": "./src/config.ts",
+    "./artifacts": "./src/artifacts.mjs",
+    "./show-video": "./src/show-video.mjs",
+    "./show-frames": "./src/show-frames.mjs"
+  },
+  "scripts": {
+    "typecheck": "tsc -p tsconfig.json --noEmit"
+  },
+  "peerDependencies": {
+    "@playwright/test": "^1.50.0"
+  }
+}
