+++ packages/e2e-record/src/index.ts
@@ -1,0 +1,11 @@
+export {
+  applyVideoZoom,
+  beginRecordedNavigation,
+  markClipStart,
+  markRecordingStart,
+  pauseForVideo,
+  recording,
+  screenshot,
+  VIDEO_ZOOM,
+} from "./recording.js";
+export { videoRecordingUse, type VideoFraming } from "./config.js";
